pyproject.toml
[project]
name = "ascetic-ddd"
version = "0.1.6"
description = "Toolkit and seedwork for a python DDD project."
authors = [
    {name = "Ivan Zakrevsky",email = "emacsway@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "psycopg-pool (>=3.3.0,<4.0.0)",
    "jsonpath2 (>=0.4.5,<0.5.0)",
    "jsonpath-rfc9535 (>=1.0.0,<2.0.0)",
    # "python-jsonpath~=2.0.2",
    "aiohttp (>=3.13.3,<4.0.0)",
    "pydash (>=8.0.5,<9.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "hypothesis (>=6.150.1,<7.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "mimesis (>=19.1.0,<20.0.0)",
    "faker (>=40.1.2,<41.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "typing_extensions (>=4.4.0)"
]

[project.optional-dependencies]
cli = [
    "jinja2 (>=3.1.0,<4.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
]

[project.scripts]
ascetic-ddd = "ascetic_ddd.cli.__main__:main"

[dependency-groups]
dev = [
    "sqlalchemy (>=2.0.45,<3.0.0)"
]
docs = [
    "sphinx (>=8.1.0,<9.0.0)",
    "sphinx-immaterial (>=0.12.4,<1.0.0)",
    "sphinx-autodoc-typehints (>=3.0.0,<4.0.0)",
    "sphinx-copybutton (>=0.5.2,<1.0.0)",
    "myst-parser (>=4.0.0,<5.0.0)",
]


[tool.poetry]
packages = [{include = "ascetic_ddd"}]
exclude = [
    "ascetic_ddd/**/tests",
    "ascetic_ddd/**/tests/**",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true
exclude = ["tests", "_build", "\\.env"]
disable_error_code = [
    "assignment",
    "override",
    "return-value",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://github.com/krew-solutions/ascetic-ddd-python"
Repository = "https://github.com/krew-solutions/ascetic-ddd-python"
Documentation = "https://krew-solutions.github.io/ascetic-ddd-python/"
"Bug Tracker" = "https://github.com/krew-solutions/ascetic-ddd-python/issues"
